backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Сервис для предоставления driven-data аналитики."
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
asyncpg = "^0.28.0"
psycopg2-binary = "^2.9.6"
SQLAlchemy = "^2.0.18"
uvicorn = "^0.22.0"
python-jose = "^3.3.0"
pydantic = "^2.0.2"
pydantic-settings = "^2.0.1"
geopy = "^2.3.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
